Rename and move primitive_pipeline_config
