+++ src/phantasm-hardware-interface/d3d12/pipeline_state.hh
@@ -3,7 +3,6 @@
 #include <clean-core/span.hh>
 
 #include <phantasm-hardware-interface/arguments.hh>
-#include <phantasm-hardware-interface/primitive_pipeline_config.hh>
 
 #include <phantasm-hardware-interface/d3d12/common/d3d12_fwd.hh>
 
@@ -14,7 +13,7 @@ namespace phi::d3d12
                                                          cc::span<D3D12_INPUT_ELEMENT_DESC const> vertex_input_layout,
                                                          const arg::framebuffer_config& framebuffer_format,
                                                          arg::graphics_shader_stages shader_stages,
-                                                         phi::primitive_pipeline_config const& config);
+                                                         graphics_pipeline_config const& config);
 
 [[nodiscard]] ID3D12PipelineState* create_compute_pipeline_state(ID3D12Device& device, ID3D12RootSignature* root_sig, const std::byte* binary_data, size_t binary_size);
 }
