+++ src/phantasm-hardware-interface/vulkan/pools/pipeline_pool.cc
@@ -13,7 +13,7 @@ phi::handle::pipeline_state phi::vk::PipelinePool::createPipelineState(phi::arg:
                                                                                        phi::arg::shader_argument_shapes shader_arg_shapes,
                                                                                        bool should_have_push_constants,
                                                                                        phi::arg::graphics_shader_stages shader_stages,
-                                                                                       const phi::primitive_pipeline_config& primitive_config)
+                                                                                       const phi::graphics_pipeline_config& primitive_config)
 {
     // Patch and reflect SPIR-V binaries
     cc::capped_vector<arg::shader_stage, 6> patched_shader_stages;
