+++ src/phantasm-hardware-interface/vulkan/pools/pipeline_pool.hh
@@ -7,7 +7,6 @@
 #include <phantasm-hardware-interface/arguments.hh>
 #include <phantasm-hardware-interface/detail/linked_pool.hh>
 #include <phantasm-hardware-interface/types.hh>
-#include <phantasm-hardware-interface/primitive_pipeline_config.hh>
 
 #include <phantasm-hardware-interface/vulkan/loader/volk.hh>
 #include <phantasm-hardware-interface/vulkan/resources/descriptor_allocator.hh>
@@ -29,7 +28,7 @@ public:
                                                              arg::shader_argument_shapes shader_arg_shapes,
                                                              bool should_have_push_constants,
                                                              arg::graphics_shader_stages shader_stages,
-                                                             phi::primitive_pipeline_config const& primitive_config);
+                                                             phi::graphics_pipeline_config const& primitive_config);
 
     [[nodiscard]] handle::pipeline_state createComputePipelineState(arg::shader_argument_shapes shader_arg_shapes,
                                                                     arg::shader_binary compute_shader,
