+++ src/phantasm-hardware-interface/vulkan/render_pass_pipeline.hh
@@ -5,13 +5,12 @@
 
 #include <phantasm-hardware-interface/arguments.hh>
 #include <phantasm-hardware-interface/commands.hh>
-#include <phantasm-hardware-interface/primitive_pipeline_config.hh>
 
 #include "loader/volk.hh"
 
 namespace phi::vk
 {
-[[nodiscard]] VkRenderPass create_render_pass(VkDevice device, const arg::framebuffer_config& framebuffer, phi::primitive_pipeline_config const& config);
+[[nodiscard]] VkRenderPass create_render_pass(VkDevice device, const arg::framebuffer_config& framebuffer, phi::graphics_pipeline_config const& config);
 
 [[nodiscard]] VkRenderPass create_render_pass(VkDevice device, const phi::cmd::begin_render_pass& begin_rp, unsigned num_samples, cc::span<const format> override_rt_formats);
 
@@ -19,7 +18,7 @@ namespace phi::vk
                                          VkRenderPass render_pass,
                                          VkPipelineLayout pipeline_layout,
                                          arg::graphics_shader_stages shaders,
-                                         phi::primitive_pipeline_config const& config,
+                                         phi::graphics_pipeline_config const& config,
                                          cc::span<VkVertexInputAttributeDescription const> vertex_attribs,
                                          uint32_t vertex_size,
                                          const arg::framebuffer_config& framebuf_config);
